+++ packages/apollo-server-core/src/index.ts
@@ -44,5 +44,5 @@ export const gql: (
 ) => DocumentNode = gqlTag;
 
 import { GraphQLScalarType } from 'graphql';
-import { GraphQLUpload as UploadScalar } from 'apollo-upload-server';
+import { GraphQLUpload as UploadScalar } from '@apollographql/apollo-upload-server';
 export const GraphQLUpload = UploadScalar as GraphQLScalarType;